restructuring cfloat mathlib and generalizing the mathlib test suite
