+++ include/sw/universal/number/cfloat/math/functions/next.hpp
