+++ include/sw/universal/number/cfloat/mathlib.hpp
@@ -6,20 +6,20 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 
-#include <universal/number/cfloat/math/classify.hpp>
-#include <universal/number/cfloat/math/complex.hpp>
-#include <universal/number/cfloat/math/error_and_gamma.hpp>
-#include <universal/number/cfloat/math/exponent.hpp>
-#include <universal/number/cfloat/math/fractional.hpp>
-#include <universal/number/cfloat/math/hyperbolic.hpp>
-#include <universal/number/cfloat/math/hypot.hpp>
-#include <universal/number/cfloat/math/logarithm.hpp>
-#include <universal/number/cfloat/math/minmax.hpp>
-#include <universal/number/cfloat/math/next.hpp>
-#include <universal/number/cfloat/math/pow.hpp>
-#include <universal/number/cfloat/math/sqrt.hpp>
-#include <universal/number/cfloat/math/trigonometry.hpp>
-#include <universal/number/cfloat/math/truncate.hpp>
+#include <universal/number/cfloat/math/functions/classify.hpp>
+#include <universal/number/cfloat/math/functions/complex.hpp>
+#include <universal/number/cfloat/math/functions/error_and_gamma.hpp>
+#include <universal/number/cfloat/math/functions/exponent.hpp>
+#include <universal/number/cfloat/math/functions/fractional.hpp>
+#include <universal/number/cfloat/math/functions/hyperbolic.hpp>
+#include <universal/number/cfloat/math/functions/hypot.hpp>
+#include <universal/number/cfloat/math/functions/logarithm.hpp>
+#include <universal/number/cfloat/math/functions/minmax.hpp>
+#include <universal/number/cfloat/math/functions/next.hpp>
+#include <universal/number/cfloat/math/functions/pow.hpp>
+#include <universal/number/cfloat/math/functions/sqrt.hpp>
+#include <universal/number/cfloat/math/functions/trigonometry.hpp>
+#include <universal/number/cfloat/math/functions/truncate.hpp>
 
 namespace sw {
     namespace universal {
